+++ src/vws/exceptions/custom_exceptions.py
@@ -5,13 +5,31 @@ or simple errors given by the cloud recognition service.
 """
 
 
-class UnknownVWSErrorPossiblyBadName(Exception):
+from requests import Response
+
+
+class OopsAnErrorOccurredPossiblyBadName(Exception):
     """
     Exception raised when VWS returns an HTML page which says "Oops, an error
     occurred".
 
     This has been seen to happen when the given name includes a bad character.
     """
+
+    def __init__(self, response: Response) -> None:
+        """
+        Args:
+            response: The response returned by Vuforia.
+        """
+        super().__init__(response.text)
+        self._response = response
+
+    @property
+    def response(self) -> Response:
+        """
+        The response returned by Vuforia which included this error.
+        """
+        return self._response
 
 
 class RequestEntityTooLarge(Exception):